requirements.txt
Flask==3.1.1
mysql-connector-python==9.4.0
python-dotenv==1.1.1
requests==2.32.4
PyJWT==2.10.1
werkzeug==3.1.3
